+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 ==========
 
+1.3.0
+------
+
+- Fix ``enctype`` attribute setting for WTForms ``MultipleFileField``.
+
 1.2.0
 ------